requirements.txt
# To ensure app dependencies are ported from your virtual environment/host machine into your container, run 'pip freeze > requirements.txt' in the terminal to overwrite this file
flask==2.0.2
gunicorn==20.1.0
flask_restful==0.3.9
flask_sqlalchemy==2.5.1
flask_jwt_extended==4.4.4
requests==2.25.1